refactor: move catalog-specific model classes to :common:catalog-spi
