+++ extensions/common/json-ld/src/main/java/org/eclipse/edc/jsonld/transformer/to/JsonObjectToDistributionTransformer.java
@@ -17,7 +17,7 @@ package org.eclipse.edc.jsonld.transformer.to;
 import jakarta.json.JsonArray;
 import jakarta.json.JsonObject;
 import jakarta.json.JsonValue;
-import org.eclipse.edc.connector.contract.spi.types.offer.Distribution;
+import org.eclipse.edc.catalog.spi.Distribution;
 import org.eclipse.edc.jsonld.transformer.AbstractJsonLdTransformer;
 import org.eclipse.edc.transform.spi.TransformerContext;
 import org.jetbrains.annotations.NotNull;
@@ -32,21 +32,21 @@ import static org.eclipse.edc.jsonld.transformer.Namespaces.DCT_SCHEMA;
 import static org.eclipse.edc.jsonld.transformer.TransformerUtil.transformString;
 
 public class JsonObjectToDistributionTransformer extends AbstractJsonLdTransformer<JsonObject, Distribution> {
-    
+
     private static final String DCAT_DISTRIBUTION_TYPE = DCAT_SCHEMA + "Distribution";
     private static final String DCAT_ACCESS_SERVICE_PROPERTY = DCAT_SCHEMA + "accessService";
     private static final String DCT_FORMAT = DCT_SCHEMA + "format";
-    
+
     public JsonObjectToDistributionTransformer() {
         super(JsonObject.class, Distribution.class);
     }
-    
+
     @Override
     public @Nullable Distribution transform(@Nullable JsonObject object, @NotNull TransformerContext context) {
         if (object == null) {
             return null;
         }
-    
+
         var type = nodeType(object, context);
         if (DCAT_DISTRIBUTION_TYPE.equals(type)) {
             var builder = Distribution.Builder.newInstance();
@@ -57,7 +57,7 @@ public class JsonObjectToDistributionTransformer extends AbstractJsonLdTransform
             return null;
         }
     }
-    
+
     private void transformProperties(String key, JsonValue value, Distribution.Builder builder, TransformerContext context) {
         if (DCAT_ACCESS_SERVICE_PROPERTY.equals(key)) {
             transformString(value, builder::dataServiceId, context);
