+++ spi/common/catalog-spi/src/main/java/org/eclipse/edc/catalog/spi/Dataset.java
@@ -12,7 +12,7 @@
  *
  */
 
-package org.eclipse.edc.connector.contract.spi.types.offer;
+package org.eclipse.edc.catalog.spi;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -25,12 +25,12 @@ import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
 import org.eclipse.edc.policy.model.Policy;
 
 public class Dataset {
-    
+
     private String id;
     private Map<String, Policy> offers = new HashMap<>();
     private List<Distribution> distributions = new ArrayList<>();
     private Map<String, Object> properties = new HashMap<>();
-    
+
     public String getId() {
         return id;
     }
@@ -38,66 +38,66 @@ public class Dataset {
     public Map<String, Policy> getOffers() {
         return offers;
     }
-    
+
     public List<Distribution> getDistributions() {
         return distributions;
     }
-    
+
     public Map<String, Object> getProperties() {
         return properties;
     }
-    
+
     @JsonIgnore
     public Object getProperty(String key) {
         return properties.get(key);
     }
-    
+
     @JsonPOJOBuilder(withPrefix = "")
     public static class Builder {
         private final Dataset dataset;
-        
+
         private Builder() {
             dataset = new Dataset();
         }
-        
+
         @JsonCreator
         public static Builder newInstance() {
             return new Builder();
         }
-        
+
         public Builder id(String id) {
             dataset.id = id;
             return this;
         }
-        
+
         public Builder offer(String offerId, Policy policy) {
             dataset.offers.put(offerId, policy);
             return this;
         }
-        
+
         public Builder distributions(List<Distribution> distributions) {
             dataset.distributions = distributions;
             return this;
         }
-        
+
         public Builder distribution(Distribution distribution) {
             dataset.distributions.add(distribution);
             return this;
         }
-        
+
         public Builder properties(Map<String, Object> properties) {
             dataset.properties = properties;
             return this;
         }
-        
+
         public Builder property(String key, Object value) {
             dataset.properties.put(key, value);
             return this;
         }
-        
+
         public Dataset build() {
             return dataset;
         }
     }
-    
+
 }
